Moved createPhotos to data
