+++ js/data/index.js
@@ -1,0 +1,4 @@
+export {default as createPhotos} from './create-photos.js';
+export {default as createPhoto} from './create-photo.js';
+export {default as generateComments} from './generate-comments.js';
+export {default as createComment} from './create-comment.js';
